move vds_with_manifest_arrays fixture up
